Refactor, throw exception on gpt4, unify experiment result structure
